+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Rename _IpregistryConfig_ option `apiUrl` to `baseUrl`.
 
 ## [3.2.0] - 2021-11-09
 ### Changed